Add type node output for full ranges
Just in case they're produced from something else, they should be output
into something that makes sense. For obvious reasons, a range with no
min and no max is just the number type, while one with both a min and a
max is output as an intersection of two half ranges.