Add range type addition check for literals/ranges
With this commit, adding (> a) + b gives you (> a + b), if b is a
numeric literal. If b is another range (> b), then you'll get (> a + b).
Adding a number or a range that is not the same "direction" will just
give you the plain number type. For example, (< a) + (> b) = number, and
(< a) + b when b is a number gives you number.